backup the rest of es results as metadata
previously we just backed up _source.dashboard which is good
for restoring a dashboard via the import interface of grafana.

however if someone wanted to implement a restore direct to ES they would
need the _id, and the rest of _source.  admittedly title and tags the important bits
are also in the _source.dashboard there is probably value in the user,
group and other data grafana may choose to store there.